requirements.txt
opencv-python>=4.8.0
replicate>=0.11.0
Pillow>=9.0.0
requests==2.32.3